Break out alert table sections into smaller components
